defining mutator and also requiring integer in product price
to avoid floating-point precision issues

+++ app/Http/Requests/Admin/Product/UpdateProductRequest.php
@@ -33,7 +33,7 @@ class UpdateProductRequest extends FormRequest
             ],
             'category_id' => ['required', 'exists:categories,id'],
             'description' => ['required', 'string'],
-            'price' => ['required', 'numeric', 'min:0'],
+            'price' => ['required', 'integer', 'min:0'],
         ];
     }
 }
